fix link display

+++ app/helpers/courses_helper.rb
@@ -1,17 +1,14 @@
 module CoursesHelper
 
-    def display_links
+    def edit_link
         if @course.teacher == current_user
-            edit_link
-            # delete_link
+            link_to "Edit Course", edit_course_path(@course)
         end
     end
 
-    def edit_link
-        link_to "Edit Course", edit_course_path(@course)
-    end
-
     def delete_link
-        link_to "Delete Course", course_path(@course), method: :delete
+        if @course.teacher == current_user
+            link_to "Delete Course", course_path(@course), method: :delete
+        end
     end
 end
